feat: swiftLint 적용 및 파일정리
- swiftLint 옵션 설정
- 불필요한 주석제거
- 컨벤션 수정
