+++ ProjectManager/.swiftlint.yml
@@ -1,0 +1,14 @@
+disabled_rules: # Default Rules에서 비활성화할 규칙
+    - trailing_whitespace
+
+
+opt_in_rules:
+  - empty_count
+  - empty_string
+
+excluded:
+  - Pods
+
+line_length: 125
+file_length: 250
+function_body_length: 15
