+++ ProjectManager/ProjectManager/Resource/Assets.xcassets/AccentColor.colorset/Contents.json
