+++ ProjectManager/ProjectManager/Resource/Assets.xcassets/AppIcon.appiconset/Contents.json
